Universal clipping code path
